Refactor webapps structure and implement prestashop installer
