+++ web/add/webapp/installer/WordpressSetup.php
@@ -1,0 +1,58 @@
+<?php
+require_once("BaseSetup.php");
+
+class WordpressSetup extends BaseSetup {
+
+    protected $appname = 'wordpress';
+    protected $config = [
+        'form' => [
+            'protocol' => [ 
+                'type' => 'select',
+                'options' => ['http','https'],
+            ],
+            'site_name' => ['type'=>'text', 'value'=>'Wordpress Blog'],
+            'site_description' => ['value'=>'Another wordpresss site'],
+            'wordpress_account_username' => ['value'=>'wpadmin'],
+            'wordpress_account_email' => 'text',
+            'wordpress_account_password' => 'password',
+            ],
+        'database' => true,
+        'url' => 'https://wordpress.org/wordpress-5.2.2.tar.gz'
+    ];
+
+    public function install(array $options) : bool {
+        parent::install($options);
+
+        $this->appcontext->runUser('v-open-fs-file',[$this->getDocRoot("wp-config-sample.php")], $result);
+
+        $distconfig = preg_replace( [
+                '/database_name_here/', '/username_here/', '/password_here/'
+            ], [
+                $this->appcontext->user() . '_' . $options['database_name'],
+                $this->appcontext->user() . '_' . $options['database_user'],
+                $options['database_password']
+            ],
+            $result->text);
+
+        while (strpos($distconfig, 'put your unique phrase here') !== false) {
+            $distconfig = preg_replace( '/put your unique phrase here/', generate_string(64), $distconfig, 1);
+        }
+
+        $tmp_configpath = $this->appcontext->saveTempFile($distconfig);
+
+        if(!$this->appcontext->runUser('v-copy-fs-file',[$tmp_configpath, $this->getDocRoot("wp-config.php")], $result)) {
+            return false;
+        }
+
+        exec("/usr/bin/curl --post301 --insecure --resolve ".$this->domain.":80:".$this->appcontext->getWebDomainIp($this->domain)." " 
+            . escapeshellarg("http://".$this->domain."/wp-admin/install.php?step=2")
+            . " -d " . escapeshellarg(
+                "weblog_title=" . rawurlencode($options['site_name'])
+            . "&user_name="      . rawurlencode($options['wordpress_account_username'])
+            . "&admin_password=" . rawurlencode($options['wordpress_account_password'])
+            . "&admin_password2=". rawurlencode($options['wordpress_account_password'])
+            . "&admin_email="    . rawurlencode($options['wordpress_account_email'])), $output, $return_var);
+
+        return ($return_var === 0);
+    }
+}
